Work in progress: Improving the quality of the TLS overview
I've improved terminology throughout the early sections and tried to
be more precise.  I probably need to add a picture or two to help.

The latter sections still need work.

I want to move the "system component" diagram that I had to the main
draft.  It would be good to have a topological map of all the pieces.