Update draft-ietf-quic-transport.md
Rewrote the PMTUD section as discussed in Issue #64:
https://github.com/quicwg/base-drafts/issues/64

Because of concern about user-space implementations controlling kernel PMTU mechanisms, and interest in PLPMTUD as an alternative to tradition ICMP PMTUD, this is a thick web of different options, MUSTs and SHOULDs. In my truest heart I would like to sweep a lot of this away and do a MUST based on around the ICMP system, but though wordy this is more likely to please everyone.