Update draft-ietf-quic-transport.md
Here is an alternate solution to he PMTUD section as discussed in Issue #64
https://github.com/quicwg/base-drafts/issues/64

I think it is much cleaner and clearer about what to do. It assumes that ICMP black holes are not a huge problem in the internet, and that no one is using a UDP socket API that prevents them from stopping various noncompliant behaviors. I much prefer this version, although some may find it too restrictive.