Use the VERSION flag to identify 0-RTT packets
This aligns the transport draft with the tls draft.

The primary reason for using this identification is that there are three types
of keys that are in concurrent use during a 0-RTT handshake: unprotected
(cleartext), 0-RTT and 1-RTT.  Identifying 0-RTT packets this way allows the
server to avoid trial decryption.