I've taken feedback from #355.
I have also added text that captures something we learned late in the HTTP/2
experience.  GOAWAY pre-emptively closes streams, but for peer-initiated
streams it might make sense to use a bigger number to allow any requests that
the peer created in the last round trip to complete.  Some implementations find
retrying quite difficult and allowing (for example) a client's last set of
requests to complete can make the shutdown properly graceful (as opposed to
toy-graceful).

The case where this matters is for a gateway to server, where the gateway can't
reasonably retry the 10,000 requests that it initiated in the last second. To
allow for that, the server can give the gateway a bit of extra leeway for those
streams when it sends a GOAWAY frame.