Limit cryptographic handshake size to 64K
Now that we have default values for the flow control offsets,
it isn't that much of a reach to use those limits for the handshake
as well.  This avoids a special case for stream 1 by imposing a
limit that no sane implementation will ever reach.  It does put a
hard limit on the size of certificates, but that limit is quite
generous.