Allow STREAM_BLOCKED on unopenable streams
There's a problem I discovered here.  We don't actually describe a transition
from idle to half-closed or closed (see discussion about half-closing and
unidirectional streams for more on this problem).  The text mentions that
RST_STREAM on an idle stream is permitted, but doesn't actually define any transitions.

I have removed the allowance for RST_STREAM, but I believe that we need that
transition.  A new issue will be opened.