Authenticate public reset with a hash
This takes ekr's design from #20 and expands on it quite a bit.  There are a
few little wrinkles that I think we might want to discuss a little.

First, this is a one-time password, so the combination of key and connection ID
can't ever be repeated for a given server instance.  Our 64-bit connection ID
space isn't really enough to provide this.  How a server moves to a new static
key during operation will be challenging; it probably needs to partition the
connection ID space.

Second, transport parameters are encrypted.  #20 made a point of having the
verifier in the clear so that intermediaries could validate the Public Reset.