Reference error codes
One interesting change here is that we don't have an error code for running out
of packet numbers.  If you think about this for even a few seconds, you will
realize that you can't send that error code because it can't go in a packet.

And that's even assuming that you manage to keep a connection alive long enough
to send that many packets.  At 1 million packets a second, that's almost 600
thousand years.  Even if we assume that you jump 2^31 packets every time you
switch between networks, and change once a minute, that's 16 thousand years.

I didn't remove the parenthetical mention of QUIC_NO_ERROR because that is
being removed in other pull requests.