Define the payload size of the initial packet rather than the packet size
The entity that creates the packet is not always aware of the existence of
header extensions.  That makes hitting the 1280 octet limit hard.  Worse, if
they are aware of extensions, they might reduce the size and then encouter a
server that enforces the limit.  That server will discard the packet.  That
makes enforcement risky.

This change places the limit on the payload.  It mandates 1232/1252 for v6/v4.

This is the uncomplicated version.  The complicated version that I didn't write
up allows a client that knows more about overheads to reduce the payload size
as long as they keep above some floor value (e.g., 1180).  We might consider
doing that later if we want to allow QUIC to operate on paths with a much
tighter MTU overhead.  Evidence suggests that we don't need this.

Closes #267