Add a description of the FNV-1a algorithm
I've added this in the TLS document, even though this would have been easier to
add to the transport document in some ways.  The current plan (if we ever clear
our backlog) is to move the details of the packet protection and key schedule
to the transport.  If that ever happens, it will be easy enough to move this
new text at the same time as the packet protection.

Two choices that I made here:

1. all packets are protected, including Public Reset - this means that the
packet construction code can have very simple logic for determining whether to
add an integrity check, namely is there an AEAD?; the cost being that public
reset is a little bigger

2. the integrity check mechanism can change in a new version - this avoids
expanding the list of things that we promise that won't change between
versions; the cost being one extra trivial check on receiving an unprotected
packet