Add a NEW_CONNECTION_ID frame
This frame is used by servers to support client unlinkability across network
migrations.

This assumes that the client is aware of the need for unlinkability.  I haven't
gone into a lot of detail on rationale, why this might be needed and the like.
That stuff does need expansion, but we don't have a lot of agreement on the
finer details.  For the moment having the basic mechanism in place will allow
us to make progress.