Rework cleartext packet descriptions
This defines a new packet type for the initial client cleartext packet, and
stateless rejects.  The remaining packets that carry cleartext messages remain
the same.  The final packet from the server is removed; instead the server
cleartext packet carries the server-selected connection ID.  The server can
propose a new connection ID in any of its cleartext packets, including the
version negotiation packet.

This will need to roll up changes from stateless rejects.  I'll merge that
branch after this commit.

Things that I might do differently: require the client to send an initial
packet after version negotiation on the basis that version negotiation is a
complete do-over.  Maybe also after a stateless reject, which would mean that
all packets containing the ClientHello would use that "initial" type.  That
would mean that only the initial type would have packet size restrictions,
which is nice.

Those alternatives give the server more options for changing connection ID as
well, but it also removes the ability for a server-side load balancer to easily
distinguish between connection IDs that the client proposes and those that the
server has provided.  Not sure which way that goes, I've opted to keep the
signal in place until we've had some more discussion.