Cover some corner cases in version negotiation
In dealing with #497, I expected that it would be possible to reject initial
packets that are too small.  I also expected that you could reject packets that
weren't of the right type.  However, in order to be able to do that we would
have to create a minimum size requirement for all QUIC versions, and make the
packet types (or at least some of them) version independent also.

That seemed unwise so instead I just described a couple of corner cases.  Short
header packets get thrown away, as do packets that are too small.

Closes #497.