Reset more completely for stateless retry and version negotiation
This is based on the discussion between editors and thus far on the list.  The
idea here is, in summary:

* Client Initial packets are always a complete reset, always contain a
ClientHello, are always padded to the minimum size, and include a randomized
packet number (except for retransmissions).  STREAM frames in these packets
always start at an offset of 0, even if the cryptographic handshake is
continued.  There can only be one of these packets.  The server can propose a
new connection ID in response to one of these packets (giving the server up to
three re-routing options).

* Server Stateless Retry packets cause the client to start over, retaining only
cryptographic handshake state.  There can only be one of these packets in a
flight.  This includes a server-proposed connection ID and an echo of the
client packet number.  This is because the server can't be expected to remember
what it chose when it sends subsequent packets.

* Version Negotiation packets also cause a reset.  This has the same rules for
connection ID and packet number as the Server Stateless Retry.

* Client Cleartext and Server Cleartext contain the other handshake messages.
These aren't really special in any way.