Force the server and client to remember transport parameters for 0-RTT
This includes Victor's recommendation to avoid the client violating server
constraints on operation.  This means that the server that accepts 0-RTT is
required to provide new transport parameters that are compatible with its
previously advertised value, or reject 0-RTT.

I made this easier on servers, by allowing them to advertise different
transport parameters that only apply to 0-RTT.  They include these values in
the TLS NewSessionTicket message, alongside the ticket that is used to resume.