Describe stateless retry in terms of a transport reset
This requires that we describe the exceptions to the process.

Minimally, this is the state necessary to ensure the integrity of the
handshake.  That is, the cryptographic handshake state continues and the client
has to remember what happened during version negotiation.  We can also point
out that keeping an RTT estimate is worthwhile, as is any congestion signals,
though these might be low value.