reword 'cumulative ack' language to avoid tcp confusion
I have had the experience of being confused by the "cumulative ack"
langauge in the current draft and then explaining it to others
confused about the same thing.

cumulative' is not used in the tcp sense where
the largest-ack is cumulative for all of the packets previous to
it. It tends to mean non-revokable here (which the draft also uses -
imo to more clarity.) While future acks can incorprate the data from
old ack frames (being in that sense cumulative) they do not have to
and in some cases will not be able to be so for space reasons.

In this case I think it is just the clarifying text that is confusing
- so less is more.