Make Related Stream ID mandatory with Offset 0
Thinking about this more, though it might be convenient to provide the field on
more than just the first STREAM frame in a stream, it is valuable to have some
certainty about the status of a stream.  Mandating the use of the field with
offset 0 ensures that a recipient has a single place that they can look for the
information.