Size limit for Version Negotiation
As discussed, the limits are largely practical, and there should be enough
incentive to avoid inflation, but it pays to be a little more precise when
defining protocols or you can get odd downstream effects.  I've chosen to limit
the size to the CI.  That's 1200 currently (or more) and should be plenty og
space for many versions.

Closes #585.