Split error code space (#722)
* Split error code space

This creates two orthogonal spaces for error codes.  Application error codes
can be used for both connection and stream errors and are under the control of
the application protocol.  Transport error codes are QUIC-controlled, but can
only terminate the connection.

To fix this, I had to add IANA considerations for error codes, plus a few extra
tweaks.  I think that the error code space could be narrowed to 16 bits after
this, if only to keep things sane.

Closes #132.

* s/Public Reset/Stateless Reset/

* Get codes consistent

* Make STOP_SENDING application-specific as well

* Split errors in HTTP too

* Refine description of HTTP codes

* Move TLS errors closer

* Match the names

* Add HTTP_NO_ERROR

* Review comments

* Jana's comments

* Small tweak at Mike's suggestion

* Reword STOP_SENDING text to avoid implication of mandatory RST_STREAM

* Better track changes on master

* Reword advice on application error codes

* Steal a code from the application

* Simply reserve the RST_STREAM code of 0.

* Application error code missed

* We really should just call this STOPPING