Enter draining after receiving a stateless reset (#868)
If a client receives a stateless reset, it makes sense for it to stick around
in the draining period so that you can absorb any further stateless reset
packets that might have been sent by the server.  The server might have
received multiple packets, so it might also generate multiple resets.