Unidirection and bidirectional streams
This is #872 with some additional exposition and one significant change.

Client-initiated streams are now EVEN.

Server initiated streams are now ODD.

This is a pretty noticeable change, because it inverts the current sign, and
might cause some disruption in implementations.  However, it makes stream 0
work within the framework better.  It wasn't clear what Ryan's intent was with
respect to numbering, but I wouldn't have noticed this until I noticed a
conflict between prose and the table.

Closes #643, #656, #720, #872, #175.

This does not close the other issues that were addressed by #643, because this
doesn't include any changes in the HTTP draft.  I'll create a follow-up PR for
that.