Make HTTP use bidirectional and unidirectional stream
The changes in #885 don't include HTTP, which leaves that draft somewhat
broken.  This corrects the omission.

There is an obvious optimization for HTTP, which is to not send the low two
bits of the Stream ID whenever it needs to identify a stream.  Right now, the
protocol only ever mentions client-initiated, bidirectional streams in frames,
so those two bits are wasted.  I'll save that change for later though.

I have chosen to take the control stream and use two unidirectional streams, which
addresses #515.

Closes #515, #240, #281.