Improve GOAWAY description
As noted by @hardie in #604, the description of GOAWAY hasn't tracked
the changes in the transport.  This updates it.  Most importantly, it says what
to do to finish a graceful shutdown, including the sending of APPLICATION_CLOSE
frames.

I added a new error code for when the server receives a GOAWAY.  I also changed
the description to be less generic and use client and server as appropriate.