Rename and consolidate packet types (#847)
* Rename and consolidate packet types

The two cleartext types are no longer needed to distinguish between client and
server.  Server packets are protected with a different key to that used for
client packets, so a server will be unable to process a packet from another
server as though it were a client packet (not that this was possible anyway
given that we have a different type for the initial packet).

Consequently, I have renamed packets to use shorter names and to better fit
their roles:

* Client Initial -> Initial
* Server Stateless Retry -> Retry (also avoids confusion with Stateless Reset)
* Server Cleartext -> Handshake
* Client Cleartext -> Handshake

I've rephrased the text in the TLS spec to match.

The only functional change here is to consolidate Server and Client Cleartext
into a single codepoint.  I did renumber the 0-RTT and 1-RTT protected packets
as a result.

Closes #846.

* Handshake, not connection setup

* Nits

* Rewrap

* Reflow, fix links, Closes #822