Renumber packet types and invert Connection ID Flag
This commit makes the following changes:
- Renumber long header packet types to count from 0x7f downwards,
  rather than from 0x01 upwards
- Renumber short header packet types to count from 0x1f downward,
  rather than from 0x01 upward
- Invert the sense of the Connection ID Flag in the short header
  packets, to become an Omit COnnection ID Flag

The purpose of the changes is to make it possible to easily demultiplex
QUIC and STUN packets running on a single UDP port, to enable peer to
peer use of QUIC. They also enable demultiplexing of WebRTC packets and
QUIC packets on a single UDP port.

Discussed in QUIC and AVTCORE working groups at IETF 100.