Move negotiated_version
This makes it possible for a client to use substantially-similar versions of
QUIC without having to regenerate a ClientHello.  That's not a huge benefit,
but given that it is no more vulnerable to downgrade attack than before, it
seems like a worthwhile thing to do.

Note that this will have the effect of violating the terms of the version
downgrade contract.  I do this knowing that this is still very much in
development and that no sane person would deploy this protocol in its current
form expecting it to be secure.  Conveniently, it appears that endpoints that
support the newer version will not be exposed to downgrade, only those that
support versions prior to the change (and even then, it's only if they can
cause the server to emit a transport parameter that looks like the current
version.  That is possible, but for QUIC version 4 only (assuming we retain the
same basic format for that long).

Closes #710.