Update key schedule
This changes the labels that we use in key derivations to be shorter.  That is
in line with similar changes TLS made, so that the number of iterations of the
underlying hash function are reduced slightly.  QUIC will still run SHA-256
like a bitcoin rig, but this should help some.

I've chosen to define the function we use rather than reuse the TLS 1.3
function.  It's still the same function, but it now uses a different root
label.  That should help avoid collisions if QUIC and TLS diverge (though that
would be unlikely).  The main benefit is in reducing the label length.

This also corrects the error in the example.