Refactor key schedule
This is an almost purely editorial change.  It consolidates the description of
QHKDF-Expand and corrects a bunch of minor things related to key derivation and
expansion.  This is the parts of #1043 that I wanted to keep.

The only non-editorial change is a change to the handshake salt used to derive
handshake keys.  That should keep the fresh.

Closes #1043.