Octets that are sent twice (#1353)
* Octets that are sent twice

This is obvious, but never stated.  We discussed checking that they stay constant and agreed that requiring a check would be crazy.  However, we never wrote anything down.  I think that a MAY is fine.

* Insist on being able to deliver in order

But don't imply an absolute requirement to do so.

It seems like an opportune moment to point out that no provision is made for partial reliability either.

* ordered stream

* Able to