Modify codepoints on decoder and request/push streams for consistency (#1472)
* Modify codepoints on decoder and request/push streams for consistency

The prefix lengths remain the same but the instruction bits now follow the rule that each instruction has at most one '1' bit.

Fixes #1471

* Swap TSS and Stream Cancellation codepoints

This is somewhat superficial, but we're changing it anyways so bear with me.  HEADER_ACK and STREAM_CANCEL are almost identical, so code processing them should be next to each other.  By swapping TSS and Stream Cancel codepoints, the processing code might be more readable.