Optimize definition IP addr in TP PreferredAddress
There are two reasons for this change.

First, having the IP address stored as an opaque string 4 to 255 octets
in length does not make it clear in what format the IP address is stored.
The choice of 4 and 16 bytes makes the format obvious.

Second, the longer of the two possibilities is 16 bytes.  There is no
reason to allow for longer values.  Only allowing 4 and 16 bytes make
the checks both stricter and simpler and implementations more efficient.