Only ack headers that have non-zero declared Largest Reference
Acknowledging headers that do not reference the dynamic table
convey no useful information to the encoder.

I took the liberty to remove the part that a header block is
"on a request or push stream."  This qualification is not required
in the context of the changed sentence.  This information is also
included in several places elsewhere in the document.