An endpoint cannot renege on advertised limits
For MAX_DATA, MAX_STREAM_DATA, and MAX_STREAM_ID equally, having a MUST
NOT on reneging isn't aligned with practice.  Some implementations
retransmit frame contents without updating the contents of these frames
and that isn't strictly a problem.

Some would have like to police this, but we seem to have consensus to
disallow that.  It creates a coupling between frame processing and
packet numbers that we don't currently have, and it seems that we don't
really want to have either.  More to the point, it prevents the sorts of
optimizations that some people really want.

So, rather than use a MUST NOT renege, observe that it is really just a
an axiom that arises as a consequence of how processing these frames
works: if the value is reduced, then ignore the value.  Whether that is
as a result of out of order arrival, or an optimization, that's OK.

Closes #1612, #447 (the latter for real this time).