Certainty is about not received
Change the text to emphasize that the requirement is that packet numbers
have never been accepted before, not that duplicates be detected.

Detecting duplicates requires an insane amount of state.  Providing
assurance that a packet has never been accepted before is easier, and
that's all we need.

Also qualify with packet number space, because the same packet number
can be sent in each of the three packet number spaces.