Separate key/secret availability from usage
As we discovered, conflating these two signals creates problems.  The
only problem we're aware of is at the server when it consumes 0-RTT, but
it's serious enough that some clarification is in order.

This makes it clear that there are two events of interest, then expands
on the particular problem scenario.

Closes #1651.