Unified ACK frame
This takes the design discussed on the list and enacts it.  This turns
out to be quite simple.  Lots of red in this change.

This treats ECT(1) as equivalent to unmarked, meaning that QUIC won't be
able to use ECT(1) without an extension or revision.  That keeps the
encoding simpler.

Closes #1439.