Tweak probing text
This text implied that there was was an expectation that the bundled NEW_CONNECTION_ID would be used.  That's not guaranteed (I expect some implementations will take the oldest unused connection ID they have and use that, for instance).  This changes it to merely acknowledge that the connection ID is unused and available as a result.