WIP: Reorganize QPACK
Added a "Compression Process Overview" section near the top, with a high level description of how to compress/decompress.  This section now encompasses a lot of what used to be "Encoding Strategies".

In the places where I added new text, I temporarily added HTML comments to indicicate so it can get a bit more detailed review.

I also deleted a few sentences/paragraphs/sections that I found to be redundant:

- Preventing Eviction Races
- "An encoder also respects..."
- "For header blocks encoded in..."
- Single Pass Encoding
- "All table updates occur on..."