Consolidate connection ID negotiation section with prior text
The text here was mostly good, and I believe that this is a good place
for that text.  However, some of the text was expository in nature, and
I moved that up to the connection ID definition section.  There is some
new stuff in there, mostly just to point forward to the negotiation
section.