I decided to reword this after reading it.  It missed a few key things.
I think that the second piece to #1786, which was to make allowances for
fatal errors arising from Initial packets won't be necessary.  I'm happy
to see a PR, of course, but the sentiment I've heard expressed is that
it is very hard to manage this, and the spec wouldn't expressly prevent
someone from trying to build a system that was resilient to
maliciously-crafted Initial packets, but we wouldn't take steps in this
document to deal with that.  As previously agreed, if you are attacked
and the attacker was able to modify Initial packets, then they can deny
you the ability to connect.

Closes #1818, #1786.