Compatible version upgrade
This introduces the design that proposed by ekr prior to the NY interim.
In this design, the client advertises the list of versions it supports
in its transport parameters.  The server is permitted to select a
"compatible" version and proceed as though the client selected that
version.

The main advantage of this approach is seamless upgrade between
compatible versions without the extra round-trip imposed by Version
Negotiation.  This will be especially useful when we move from
pre-release versions (0xff0000xx) to the final version.

Closes #1773, #1755.