Define stream limits as counts
The problems we've been having with stream limits turned out to be
intractable without a larger shift.  This aligns stream limits with flow
control limits.  To do that, it takes two frame types, both for
increasing the limit and for signaling blocking.

Note that these frame types will be renumbered afterwards.  I realize
that this is a terrible arrangement, but it's temporary.  Ideally all
this limiting stuff is in a contiguous block.

Closes #1850.