Thwart downgrade attacks between incompatible versions
This complicates things slightly more.  The client has to order its
supported versions strictly (and respect that ordering in its selection
of versions).  The server has to validate things more carefully.

I have an algorithm for validation in comments, I'm interested in
whether that is worth putting in the draft.