Stop reserving application error codes
We really don't need to do this.  As observed, the error code in
RST_STREAM really doesn't matter once you have requested that it stop.

I know that this wasn't a universally-held opinion, but we don't need to
stomp on values in the application protocol, so we shouldn't.  (Igor
suggested a new frame type, and that is maybe cleaner, but it's
wasteful when it won't be used, so I think that this is better.)

Closes #1804.