Flow control hazards
This describes some of the risks entailed by flow control.  I'm starting
to think that this sort of language is best moved to the applicability
statement, but I'd like other opinions on that.

I'm also not completely happy with how this turned out.  This is just
the first rough cut.  Concrete suggestions for how to improve this would
be greatly appreciated.

Closes #1432.